Enum bevy_asset::AssetId

pub enum AssetId<A: Asset> {
    Index {
        index: AssetIndex,
        marker: PhantomData<fn() -> A>,
    },
    Uuid {
        uuid: Uuid,
    },
}

A unique runtime-only identifier for an Asset. This is cheap to Copy/Clone and is not directly tied to the lifetime of the Asset. This means it can point to an Asset that no longer exists.

For an identifier tied to the lifetime of an asset, see Handle.

For an “untyped” / “generic-less” id, see UntypedAssetId.

Variants

Index

A small / efficient runtime identifier that can be used to efficiently look up an asset stored in Assets. This is the “default” identifier used for assets. The alternative(s) (ex: AssetId::Uuid) will only be used if assets are explicitly registered that way.

Fields

index: AssetIndex

marker: PhantomData<fn() -> A>

Uuid

A stable-across-runs / const asset identifier. This will only be used if an asset is explicitly registered in Assets with one.

Fields

uuid: Uuid

Implementations

impl<A: Asset> AssetId<A>

pub const DEFAULT_UUID: Uuid = _

The uuid for the default AssetId. It is valid to assign a value to this in Assets and by convention (where appropriate) assets should support this pattern.

pub const INVALID_UUID: Uuid = _

This asset id should never be valid. Assigning a value to this in Assets will produce undefined behavior, so don’t do it!

pub const fn invalid() -> Self

Returns an AssetId with Self::INVALID_UUID, which should never be assigned to.

pub fn untyped(self) -> UntypedAssetId

Converts this to an “untyped” / “generic-less” Asset identifier that stores the type information inside the UntypedAssetId.