Enum bevy_asset::UntypedHandle

pub enum UntypedHandle {
    Strong(Arc<StrongHandle>),
    Weak(UntypedAssetId),
}

An untyped variant of Handle, which internally stores the Asset type information at runtime as a TypeId instead of encoding it in the compile-time type. This allows handles across Asset types to be stored together and compared.

See Handle for more information.

Variants

Strong(Arc<StrongHandle>)

Weak(UntypedAssetId)

Implementations

impl UntypedHandle

pub fn id(&self) -> UntypedAssetId

Returns the UntypedAssetId for the referenced asset.

pub fn path(&self) -> Option<&AssetPath<'static>>

Returns the path if this is (1) a strong handle and (2) the asset has a path

pub fn clone_weak(&self) -> UntypedHandle

Creates an UntypedHandle::Weak clone of this UntypedHandle, which will not keep the referenced Asset alive.

pub fn type_id(&self) -> TypeId

Returns the TypeId of the referenced Asset.

pub fn typed_unchecked<A: Asset>(self) -> Handle<A>

Converts to a typed Handle. This will not check if the target Handle type matches.

pub fn typed_debug_checked<A: Asset>(self) -> Handle<A>

Converts to a typed Handle. This will check the type when compiled with debug asserts, but it will not check if the target Handle type matches in release builds. Use this as an optimization when you want some degree of validation at dev-time, but you are also very certain that the type actually matches.

pub fn typed<A: Asset>(self) -> Handle<A>

Converts to a typed Handle. This will panic if the internal TypeId does not match the given asset type A

pub fn try_typed<A: Asset>( self ) -> Result<Handle<A>, UntypedAssetConversionError>

Converts to a typed Handle. This will panic if the internal TypeId does not match the given asset type A

pub fn meta_transform(&self) -> Option<&MetaTransform>

The “meta transform” for the strong handle. This will only be Some if the handle is strong and there is a meta transform associated with it.