Struct bevy_asset::io::file::FileAssetReader

pub struct FileAssetReader { /* private fields */ }

I/O implementation for the local filesystem.

This asset I/O is fully featured but it’s not available on android and wasm targets.

Implementations

impl FileAssetReader

pub fn new<P: AsRef<Path>>(path: P) -> Self

Creates a new FileAssetIo at a path relative to the executable’s directory, optionally watching for changes.

See get_base_path below.

pub fn get_base_path() -> PathBuf

Returns the base path of the assets directory, which is normally the executable’s parent directory.

If the CARGO_MANIFEST_DIR environment variable is set, then its value will be used instead. It’s set by cargo when running with cargo run.

pub fn root_path(&self) -> &PathBuf

Returns the root directory where assets are loaded from.

See get_base_path.