Trait bevy_asset::io::AssetReader

pub trait AssetReader: Send + Sync + 'static {
    // Required methods
    fn read<'a>(
        &'a self,
        path: &'a Path
    ) -> BoxedFuture<'a, Result<Box<Reader<'a>>, AssetReaderError>>;
    fn read_meta<'a>(
        &'a self,
        path: &'a Path
    ) -> BoxedFuture<'a, Result<Box<Reader<'a>>, AssetReaderError>>;
    fn read_directory<'a>(
        &'a self,
        path: &'a Path
    ) -> BoxedFuture<'a, Result<Box<PathStream>, AssetReaderError>>;
    fn is_directory<'a>(
        &'a self,
        path: &'a Path
    ) -> BoxedFuture<'a, Result<bool, AssetReaderError>>;

    // Provided method
    fn read_meta_bytes<'a>(
        &'a self,
        path: &'a Path
    ) -> BoxedFuture<'a, Result<Vec<u8>, AssetReaderError>> { ... }
}

Performs read operations on an asset storage. AssetReader exposes a “virtual filesystem” API, where asset bytes and asset metadata bytes are both stored and accessible for a given path.

Also see AssetWriter.

Required Methods

fn read<'a>( &'a self, path: &'a Path ) -> BoxedFuture<'a, Result<Box<Reader<'a>>, AssetReaderError>>

Returns a future to load the full file data at the provided path.

fn read_meta<'a>( &'a self, path: &'a Path ) -> BoxedFuture<'a, Result<Box<Reader<'a>>, AssetReaderError>>

Returns a future to load the full file data at the provided path.

fn read_directory<'a>( &'a self, path: &'a Path ) -> BoxedFuture<'a, Result<Box<PathStream>, AssetReaderError>>

Returns an iterator of directory entry names at the provided path.

fn is_directory<'a>( &'a self, path: &'a Path ) -> BoxedFuture<'a, Result<bool, AssetReaderError>>

Returns true if the provided path points to a directory.

Provided Methods

fn read_meta_bytes<'a>( &'a self, path: &'a Path ) -> BoxedFuture<'a, Result<Vec<u8>, AssetReaderError>>

Reads asset metadata bytes at the given path into a Vec<u8>. This is a convenience function that wraps AssetReader::read_meta by default.