Struct bevy_asset::meta::AssetMeta

pub struct AssetMeta<L: AssetLoader, P: Process> {
    pub meta_format_version: String,
    pub processed_info: Option<ProcessedInfo>,
    pub asset: AssetAction<L::Settings, P::Settings>,
}

Asset metadata that informs how an Asset should be handled by the asset system.

L is the AssetLoader (if one is configured) for the AssetAction. This can be () if it is not required. P is the Process processor, if one is configured for the AssetAction. This can be () if it is not required.

Fields
meta_format_version: String

The version of the meta format being used. This will change whenever a breaking change is made to the meta format.

processed_info: Option<ProcessedInfo>

Information produced by the AssetProcessor after processing this asset. This will only exist alongside processed versions of assets. You should not manually set it in your asset source files.

asset: AssetAction<L::Settings, P::Settings>

How to handle this asset in the asset system. See AssetAction.

Implementations

impl<L: AssetLoader, P: Process> AssetMeta<L, P>

pub fn new(asset: AssetAction<L::Settings, P::Settings>) -> Self

pub fn deserialize(bytes: &[u8]) -> Result<Self, DeserializeMetaError>

Deserializes the given serialized byte representation of the asset meta.