Trait bevy_asset::meta::AssetMetaDyn

pub trait AssetMetaDyn: Downcast + Send + Sync {
    // Required methods
    fn loader_settings(&self) -> Option<&dyn Settings>;
    fn loader_settings_mut(&mut self) -> Option<&mut dyn Settings>;
    fn serialize(&self) -> Vec<u8>;
    fn processed_info(&self) -> &Option<ProcessedInfo>;
    fn processed_info_mut(&mut self) -> &mut Option<ProcessedInfo>;
}

A dynamic type-erased counterpart to AssetMeta that enables passing around and interacting with AssetMeta without knowing its type.

Required Methods

fn loader_settings(&self) -> Option<&dyn Settings>

Returns a reference to the AssetLoader settings, if they exist.

fn loader_settings_mut(&mut self) -> Option<&mut dyn Settings>

Returns a mutable reference to the AssetLoader settings, if they exist.

fn serialize(&self) -> Vec<u8>

Serializes the internal AssetMeta.

fn processed_info(&self) -> &Option<ProcessedInfo>

Returns a reference to the ProcessedInfo if it exists.

fn processed_info_mut(&mut self) -> &mut Option<ProcessedInfo>

Returns a mutable reference to the ProcessedInfo if it exists.

Implementations

impl dyn AssetMetaDyn

pub fn is<__T: AssetMetaDyn>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: AssetMetaDyn>( self: Box<Self> ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: AssetMetaDyn>( self: Rc<Self> ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: AssetMetaDyn>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: AssetMetaDyn>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.