bevy_asset::meta
Trait Settings

pub trait Settings:
    Downcast
    + Send
    + Sync
    + 'static { }

Settings used by the asset system, such as by AssetLoader, Process, and AssetSaver

Implementations

impl dyn Settings

pub fn is<__T: Settings>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Settings>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Settings>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Settings>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Settings>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.