Struct bevy_asset::processor::AssetProcessor

pub struct AssetProcessor { /* private fields */ }

A “background” asset processor that reads asset values from a source AssetSource (which corresponds to an AssetReader / AssetWriter pair), processes them in some way, and writes them to a destination AssetSource.

This will create .meta files (a human-editable serialized form of AssetMeta) in the source AssetSource for assets that that can be loaded and/or processed. This enables developers to configure how each asset should be loaded and/or processed.

AssetProcessor can be run in the background while a Bevy App is running. Changes to assets will be automatically detected and hot-reloaded.

Assets will only be re-processed if they have been changed. A hash of each asset source is stored in the metadata of the processed version of the asset, which is used to determine if the asset source has actually changed.

A ProcessorTransactionLog is produced, which uses “write-ahead logging” to make the AssetProcessor crash and failure resistant. If a failed/unfinished transaction from a previous run is detected, the affected asset(s) will be re-processed.

AssetProcessor can be cloned. It is backed by an Arc so clones will share state. Clones can be freely used in parallel.

Implementations

impl AssetProcessor

pub fn new(source: &mut AssetSourceBuilders) -> Self

Creates a new AssetProcessor instance.

pub fn server(&self) -> &AssetServer

The “internal” AssetServer used by the AssetProcessor. This is separate from the asset processor used by the main App. It has different processor-specific configuration and a different ID space.

pub async fn get_state(&self) -> ProcessorState

Retrieves the current ProcessorState

pub fn get_source<'a, 'b>( &'a self, id: impl Into<AssetSourceId<'b>> ) -> Result<&'a AssetSource, MissingAssetSourceError>

Retrieves the AssetSource for this processor

pub fn sources(&self) -> &AssetSources

pub fn start(_processor: Res<'_, Self>)

Starts the processor in a background thread.

pub async fn listen_for_source_change_events(&self)

Listens for changes to assets in the source AssetSource and update state accordingly.

pub fn register_processor<P: Process>(&self, processor: P)

Register a new asset processor.

pub fn set_default_processor<P: Process>(&self, extension: &str)

Set the default processor for the given extension. Make sure P is registered with AssetProcessor::register_processor.

pub fn get_default_processor( &self, extension: &str ) -> Option<Arc<dyn ErasedProcessor>>

Returns the default processor for the given extension, if it exists.

pub fn get_processor( &self, processor_type_name: &str ) -> Option<Arc<dyn ErasedProcessor>>

Returns the processor with the given processor_type_name, if it exists.