Module bevy_asset::saver

Structs

SavedAsset

An Asset (and any labeled “sub assets”) intended to be saved.

Traits

AssetSaver

Saves an Asset of a given AssetSaver::Asset type. AssetSaver::OutputLoader will then be used to load the saved asset in the final deployed application. The saver should produce asset bytes in a format that AssetSaver::OutputLoader can read.

ErasedAssetSaver

A type-erased dynamic variant of AssetSaver that allows callers to save assets without knowing the actual type of the AssetSaver.