bevy_asset
Trait AssetContainer

pub trait AssetContainer:
    Downcast
    + Any
    + Send
    + Sync
    + 'static {
    // Required methods
    fn insert(self: Box<Self>, id: UntypedAssetId, world: &mut World);
    fn asset_type_name(&self) -> &'static str;
}

A type erased container for an Asset value that is capable of inserting the Asset into a World’s Assets collection.

Required Methods

fn insert(self: Box<Self>, id: UntypedAssetId, world: &mut World)

fn asset_type_name(&self) -> &'static str

Implementations

impl dyn AssetContainer

pub fn is<__T: AssetContainer>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: AssetContainer>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: AssetContainer>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: AssetContainer>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: AssetContainer>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.