Trait bevy_asset::AssetDynamic

pub trait AssetDynamic: Downcast + TypeUuidDynamic + Send + Sync + 'static { }

Implementations

impl dyn AssetDynamic

pub fn is<__T: AssetDynamic>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: AssetDynamic>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn't.

pub fn downcast_rc<__T: AssetDynamic>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn't.

pub fn downcast_ref<__T: AssetDynamic>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn't.

pub fn downcast_mut<__T: AssetDynamic>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn't.