Trait bevy_asset::AssetLoader

pub trait AssetLoader: Send + Sync + 'static {
    type Asset: Asset;
    type Settings: Settings + Default + Serialize + for<'a> Deserialize<'a>;
    type Error: Into<Box<dyn Error + Send + Sync + 'static>>;

    // Required method
    fn load<'a>(
        &'a self,
        reader: &'a mut Reader<'_>,
        settings: &'a Self::Settings,
        load_context: &'a mut LoadContext<'_>
    ) -> impl ConditionalSendFuture<Output = Result<Self::Asset, Self::Error>>;

    // Provided method
    fn extensions(&self) -> &[&str] { ... }
}

Loads an Asset from a given byte Reader. This can accept AssetLoader::Settings, which configure how the Asset should be loaded.

Required Associated Types

type Asset: Asset

The top level Asset loaded by this AssetLoader.

type Settings: Settings + Default + Serialize + for<'a> Deserialize<'a>

The settings type used by this AssetLoader.

type Error: Into<Box<dyn Error + Send + Sync + 'static>>

The type of error which could be encountered by this loader.

Required Methods

fn load<'a>( &'a self, reader: &'a mut Reader<'_>, settings: &'a Self::Settings, load_context: &'a mut LoadContext<'_> ) -> impl ConditionalSendFuture<Output = Result<Self::Asset, Self::Error>>

Asynchronously loads AssetLoader::Asset (and any other labeled assets) from the bytes provided by Reader.

Provided Methods

fn extensions(&self) -> &[&str]

Returns a list of extensions supported by this AssetLoader, without the preceding dot. Note that users of this AssetLoader may choose to load files with a non-matching extension.

Object Safety

This trait is not object safe.