Struct bevy_asset::transformer::TransformedAsset

pub struct TransformedAsset<A: Asset> { /* private fields */ }

An Asset (and any “sub assets”) intended to be transformed

Implementations

impl<A: Asset> TransformedAsset<A>

pub fn from_loaded(asset: ErasedLoadedAsset) -> Option<Self>

Creates a new TransformedAsset from asset if its internal value matches A.

pub fn replace_asset<B: Asset>(self, asset: B) -> TransformedAsset<B>

Creates a new TransformedAsset from asset, transferring the labeled_assets from this TransformedAsset to the new one

pub fn take_labeled_assets<B: Asset>( &mut self, labeled_source: TransformedAsset<B> )

Takes the labeled assets from labeled_source and places them in this TransformedAsset

pub fn get(&self) -> &A

Retrieves the value of this asset.

pub fn get_mut(&mut self) -> &mut A

Mutably retrieves the value of this asset.

pub fn get_labeled<B: Asset, Q>( &mut self, label: &Q ) -> Option<TransformedSubAsset<'_, B>>
where CowArc<'static, str>: Borrow<Q>, Q: ?Sized + Hash + Eq,

Returns the labeled asset, if it exists and matches this type.

pub fn get_erased_labeled<Q>(&self, label: &Q) -> Option<&ErasedLoadedAsset>
where CowArc<'static, str>: Borrow<Q>, Q: ?Sized + Hash + Eq,

Returns the type-erased labeled asset, if it exists and matches this type.

pub fn get_untyped_handle<Q>(&self, label: &Q) -> Option<UntypedHandle>
where CowArc<'static, str>: Borrow<Q>, Q: ?Sized + Hash + Eq,

Returns the UntypedHandle of the labeled asset with the provided ‘label’, if it exists.

pub fn get_handle<Q, B: Asset>(&self, label: &Q) -> Option<Handle<B>>
where CowArc<'static, str>: Borrow<Q>, Q: ?Sized + Hash + Eq,

Returns the Handle of the labeled asset with the provided ‘label’, if it exists and is an asset of type B

pub fn insert_labeled( &mut self, label: impl Into<CowArc<'static, str>>, handle: impl Into<UntypedHandle>, asset: impl Into<ErasedLoadedAsset> )

Adds asset as a labeled sub asset using label and handle

pub fn iter_labels(&self) -> impl Iterator<Item = &str>

Iterate over all labels for “labeled assets” in the loaded asset