Crate bevy_asset

Modules

diagnostic

prelude

Macros

load_internal_asset

Structs

AssetLifecycleChannel

A channel to send and receive AssetResults

AssetMeta

AssetPath

AssetPathId

AssetPlugin

Adds support for Assets to an App. Assets are typed collections with change tracking, which are added as App Resources. Examples of assets: textures, sounds, 3d models, maps, scenes

AssetResult

The result of loading an asset of type T

AssetServer

Loads assets from the filesystem on background threads

AssetServerInternal

AssetServerSettings

Assets

Stores Assets of a given type and tracks changes to them.

FileAssetIo

Handle

A handle into a specific Asset of type T

HandleUntyped

A non-generic version of Handle

LabelId

LoadContext

LoadedAsset

SourceInfo

Info about a specific asset, such as its path and its current load state

SourceMeta

SourcePathId

Enums

AssetEvent

Events that happen on assets of type T

AssetIoError

Errors that occur while loading assets

AssetLifecycleEvent

AssetServerError

Errors that occur while loading assets with an AssetServer

AssetStage

The names of asset stages in an App Schedule

HandleId

A unique, stable asset id

LoadState

The load state of an asset

Traits

AddAsset

App extension methods for adding new asset types

Asset

AssetDynamic

AssetIo

Handles load requests from an AssetServer

AssetLifecycle

AssetLoader

A loader for an asset source

Functions

create_platform_default_asset_io

Create an instance of the platform default AssetIo

free_unused_assets_system

update_asset_storage_system

Updates the Assets collection according to the changes queued up by AssetServer.

Type Definitions

BoxedFuture