Struct bevy_asset::io::AssetSource

pub struct AssetSource { /* private fields */ }

A collection of unprocessed and processed AssetReader, AssetWriter, and AssetWatcher instances for a specific asset source, identified by an AssetSourceId.

Implementations

impl AssetSource

pub fn build() -> AssetSourceBuilder

Starts building a new AssetSource.

pub fn id(&self) -> AssetSourceId<'static>

Returns this source’s id.

pub fn reader(&self) -> &dyn ErasedAssetReader

Return’s this source’s unprocessed AssetReader.

pub fn writer(&self) -> Result<&dyn ErasedAssetWriter, MissingAssetWriterError>

Return’s this source’s unprocessed AssetWriter, if it exists.

pub fn processed_reader( &self ) -> Result<&dyn ErasedAssetReader, MissingProcessedAssetReaderError>

Return’s this source’s processed AssetReader, if it exists.

pub fn processed_writer( &self ) -> Result<&dyn ErasedAssetWriter, MissingProcessedAssetWriterError>

Return’s this source’s processed AssetWriter, if it exists.

pub fn event_receiver(&self) -> Option<&Receiver<AssetSourceEvent>>

Return’s this source’s unprocessed event receiver, if the source is currently watching for changes.

pub fn processed_event_receiver(&self) -> Option<&Receiver<AssetSourceEvent>>

Return’s this source’s processed event receiver, if the source is currently watching for changes.

pub fn should_process(&self) -> bool

Returns true if the assets in this source should be processed.

pub fn get_default_reader( _path: String ) -> impl FnMut() -> Box<dyn ErasedAssetReader> + Send + Sync

Returns a builder function for this platform’s default AssetReader. path is the relative path to the asset root.

pub fn get_default_writer( _path: String ) -> impl FnMut(bool) -> Option<Box<dyn ErasedAssetWriter>> + Send + Sync

Returns a builder function for this platform’s default AssetWriter. path is the relative path to the asset root. This will return None if this platform does not support writing assets by default.

pub fn get_default_watch_warning() -> &'static str

Returns the default non-existent AssetWatcher warning for the current platform.

pub fn get_default_watcher( path: String, file_debounce_wait_time: Duration ) -> impl FnMut(Sender<AssetSourceEvent>) -> Option<Box<dyn AssetWatcher>> + Send + Sync

Returns a builder function for this platform’s default AssetWatcher. path is the relative path to the asset root. This will return None if this platform does not support watching assets by default. file_debounce_time is the amount of time to wait (and debounce duplicate events) before returning an event. Higher durations reduce duplicates but increase the amount of time before a change event is processed. If the duration is set too low, some systems might surface events before their filesystem has the changes.

pub fn gate_on_processor(&mut self, processor_data: Arc<AssetProcessorData>)

This will cause processed AssetReader futures (such as AssetReader::read) to wait until the AssetProcessor has finished processing the requested asset.