Module bevy_asset::processor

Structs

AssetProcessor

A “background” asset processor that reads asset values from a source AssetSource (which corresponds to an AssetReader / AssetWriter pair), processes them in some way, and writes them to a destination AssetSource.

AssetProcessorData

LoadAndSave

A flexible Process implementation that loads the source Asset using the L AssetLoader, then saves that L asset using the S AssetSaver.

LoadAndSaveSettings

Settings for the LoadAndSave Process::Settings implementation.

ProcessContext

Provides scoped data access to the AssetProcessor. This must only expose processor data that is represented in the asset’s hash.

ProcessorAssetInfos

The “current” in memory view of the asset space. This is “eventually consistent”. It does not directly represent the state of assets in storage, but rather a valid historical view that will gradually become more consistent as events are processed.

ProcessorTransactionLog

A “write ahead” logger that helps ensure asset importing is transactional. Prior to processing an asset, we write to the log to indicate it has started After processing an asset, we write to the log to indicate it has finished. On startup, the log can be read to determine if any transactions were incomplete.

WriteLogError

An error that occurs when writing to the ProcessorTransactionLog fails.

Enums

InitializeError

An error that occurs when initializing the AssetProcessor.

LogEntryError

An error that occurs when validating individual ProcessorTransactionLog entries.

ProcessError

An error that is encountered during Process::process.

ProcessResult

The (successful) result of processing an asset

ProcessStatus

The final status of processing an asset

ProcessorState

The current state of the AssetProcessor.

ReadLogError

An error that occurs when reading from the ProcessorTransactionLog fails.

ValidateLogError

An error that occurs when validating the ProcessorTransactionLog fails.

Traits

ErasedProcessor

A type-erased variant of Process that enables interacting with processor implementations without knowing their type.

Process

Asset “processor” logic that reads input asset bytes (stored on ProcessContext), processes the value in some way, and then writes the final processed bytes with Writer. The resulting bytes must be loadable with the given Process::OutputLoader.