Struct bevy_asset::Assets

pub struct Assets<A: Asset> { /* private fields */ }

Stores Asset values identified by their AssetId.

Assets identified by AssetId::Index will be stored in a “dense” vec-like storage. This is more efficient, but it means that the assets can only be identified at runtime. This is the default behavior.

Assets identified by AssetId::Uuid will be stored in a hashmap. This is less efficient, but it means that the assets can be referenced at compile time.

This tracks (and queues) AssetEvent events whenever changes to the collection occur.

Implementations

impl<A: Asset> Assets<A>

pub fn get_handle_provider(&self) -> AssetHandleProvider

Retrieves an AssetHandleProvider capable of reserving new Handle values for assets that will be stored in this collection.

pub fn reserve_handle(&self) -> Handle<A>

Reserves a new Handle for an asset that will be stored in this collection.

pub fn insert(&mut self, id: impl Into<AssetId<A>>, asset: A)

Inserts the given asset, identified by the given id. If an asset already exists for id, it will be replaced.

pub fn get_or_insert_with( &mut self, id: impl Into<AssetId<A>>, insert_fn: impl FnOnce() -> A ) -> &mut A

Retrieves an Asset stored for the given id if it exists. If it does not exist, it will be inserted using insert_fn.

pub fn contains(&self, id: impl Into<AssetId<A>>) -> bool

Returns true if the id exists in this collection. Otherwise it returns false.

pub fn add(&mut self, asset: impl Into<A>) -> Handle<A>

Adds the given asset and allocates a new strong Handle for it.

pub fn get_strong_handle(&mut self, id: AssetId<A>) -> Option<Handle<A>>

Upgrade an AssetId into a strong Handle that will prevent asset drop.

Returns None if the provided id is not part of this Assets collection. For example, it may have been dropped earlier.

pub fn get(&self, id: impl Into<AssetId<A>>) -> Option<&A>

Retrieves a reference to the Asset with the given id, if it exists. Note that this supports anything that implements Into<AssetId<A>>, which includes Handle and AssetId.

pub fn get_mut(&mut self, id: impl Into<AssetId<A>>) -> Option<&mut A>

Retrieves a mutable reference to the Asset with the given id, if it exists. Note that this supports anything that implements Into<AssetId<A>>, which includes Handle and AssetId.

pub fn remove(&mut self, id: impl Into<AssetId<A>>) -> Option<A>

Removes (and returns) the Asset with the given id, if it exists. Note that this supports anything that implements Into<AssetId<A>>, which includes Handle and AssetId.

pub fn remove_untracked(&mut self, id: impl Into<AssetId<A>>) -> Option<A>

Removes (and returns) the Asset with the given id, if it exists. This skips emitting AssetEvent::Removed. Note that this supports anything that implements Into<AssetId<A>>, which includes Handle and AssetId.

pub fn is_empty(&self) -> bool

Returns true if there are no assets in this collection.

pub fn len(&self) -> usize

Returns the number of assets currently stored in the collection.

pub fn ids(&self) -> impl Iterator<Item = AssetId<A>> + '_

Returns an iterator over the AssetId of every Asset stored in this collection.

pub fn iter(&self) -> impl Iterator<Item = (AssetId<A>, &A)>

Returns an iterator over the AssetId and Asset ref of every asset in this collection.

pub fn iter_mut(&mut self) -> AssetsMutIterator<'_, A>

Returns an iterator over the AssetId and mutable Asset ref of every asset in this collection.

pub fn track_assets( assets: ResMut<'_, Self>, asset_server: Res<'_, AssetServer> )

A system that synchronizes the state of assets in this collection with the AssetServer. This manages Handle drop events.

pub fn asset_events( assets: ResMut<'_, Self>, events: EventWriter<'_, AssetEvent<A>> )

A system that applies accumulated asset change events to the Events resource.