Struct bevy_asset::Assets

pub struct Assets<T: Asset> { /* private fields */ }

Stores Assets of a given type and tracks changes to them.

Each asset is mapped by a unique HandleId, allowing any Handle with the same HandleId to access it. These assets remain loaded for as long as a Strong handle to that asset exists.

To store a reference to an asset without forcing it to stay loaded, you can use a Weak handle. To make a Weak handle a Strong one, use Assets::get_handle or pass the Assets collection into the handle’s make_strong method.

Remember, if there are no Strong handles for an asset (i.e. they have all been dropped), the asset will unload. Make sure you always have a Strong handle when you want to keep an asset loaded!

Implementations

impl<T: Asset> Assets<T>

pub fn add(&mut self, asset: T) -> Handle<T>

Adds an asset to the collection, returning a Strong handle to that asset.

Events

AssetEvent::Created

pub fn set<H: Into<HandleId>>(&mut self, handle: H, asset: T) -> Handle<T>

Add/modify the asset pointed to by the given handle.

Unless there exists another Strong handle for this asset, it’s advised to use the returned Strong handle. Not doing so may result in the unexpected release of the asset.

See set_untracked for more info.

pub fn set_untracked<H: Into<HandleId>>(&mut self, handle: H, asset: T)

Add/modify the asset pointed to by the given handle.

If an asset already exists with the given HandleId, it will be modified. Otherwise the new asset will be inserted.

Events

AssetEvent::Created: Sent if the asset did not yet exist with the given handle.

AssetEvent::Modified: Sent if the asset with given handle already existed.

pub fn get(&self, handle: &Handle<T>) -> Option<&T>

Gets the asset for the given handle.

This is the main method for accessing asset data from an Assets collection. If you need mutable access to the asset, use get_mut.

pub fn contains(&self, handle: &Handle<T>) -> bool

Checks if an asset exists for the given handle

pub fn get_mut(&mut self, handle: &Handle<T>) -> Option<&mut T>

Get mutable access to the asset for the given handle.

This is the main method for mutably accessing asset data from an Assets collection. If you do not need mutable access to the asset, you may also use get.

pub fn get_handle<H: Into<HandleId>>(&self, handle: H) -> Handle<T>

Gets a Strong handle pointing to the same asset as the given one.

pub fn get_or_insert_with<H: Into<HandleId>>( &mut self, handle: H, insert_fn: impl FnOnce() -> T ) -> &mut T

Gets mutable access to an asset for the given handle, inserting a new value if none exists.

Events

AssetEvent::Created: Sent if the asset did not yet exist with the given handle.

pub fn iter(&self) -> impl Iterator<Item = (HandleId, &T)>

Gets an iterator over all assets in the collection.

pub fn iter_mut(&mut self) -> impl Iterator<Item = (HandleId, &mut T)>

Gets a mutable iterator over all assets in the collection.

pub fn ids(&self) -> impl Iterator<Item = HandleId> + '_

Gets an iterator over all HandleId’s in the collection.

pub fn remove<H: Into<HandleId>>(&mut self, handle: H) -> Option<T>

Removes an asset for the given handle.

The asset is returned if it existed in the collection, otherwise None.

Events

AssetEvent::Removed

pub fn clear(&mut self)

Clears the inner asset map, removing all key-value pairs.

Keeps the allocated memory for reuse.

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted into the assets.

The collection may reserve more space to avoid frequent reallocations.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the asset map as much as possible.

It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

pub fn asset_event_system( events: EventWriter<'_, AssetEvent<T>>, assets: ResMut<'_, Assets<T>> )

A system that creates AssetEvents at the end of the frame based on changes in the asset storage.

pub fn len(&self) -> usize

Gets the number of assets in the collection.

pub fn is_empty(&self) -> bool

Returns true if there are no stored assets.