Struct bevy_asset::LoadContext

pub struct LoadContext<'a> { /* private fields */ }

An asynchronous context where an Asset is processed.

The load context is created by the AssetServer to process an asset source after loading its contents into memory. It is then passed to the appropriate AssetLoader based on the file extension of the asset’s path.

An asset source can define one or more assets from a single source path. The main asset is set using LoadContext::set_default_asset and sub-assets are defined with LoadContext::set_labeled_asset.

Implementations

impl<'a> LoadContext<'a>

pub fn path(&self) -> &Path

Gets the source path for this load context.

pub fn has_labeled_asset(&self, label: &str) -> bool

Returns true if the load context contains an asset with the specified label.

pub fn set_default_asset<T: Asset>(&mut self, asset: LoadedAsset<T>)

Sets the primary asset loaded from the asset source.

pub fn set_labeled_asset<T: Asset>( &mut self, label: &str, asset: LoadedAsset<T> ) -> Handle<T>

Sets a secondary asset loaded from the asset source.

pub fn get_handle<I: Into<HandleId>, T: Asset>(&self, id: I) -> Handle<T>

Gets a strong handle to an asset of type T from its id.

pub fn get_handle_untyped<I: Into<HandleId>>(&self, id: I) -> HandleUntyped

Gets an untyped strong handle for an asset with the provided id.

pub async fn read_asset_bytes<P: AsRef<Path>>( &self, path: P ) -> Result<Vec<u8>, AssetIoError>

Reads the contents of the file at the specified path through the AssetIo associated with this context.

pub fn get_asset_metas(&self) -> Vec<AssetMeta>

Generates metadata for the assets managed by this load context.

pub fn asset_io(&self) -> &dyn AssetIo

Gets the asset I/O associated with this load context.