Trait bevy_asset::AssetLoader

pub trait AssetLoader: Send + Sync + 'static {
    type Asset: Asset;
    type Settings: Settings + Default + Serialize + for<'a> Deserialize<'a>;
    type Error: Into<Box<dyn Error + Send + Sync + 'static>>;

    // Required methods
    fn load<'a>(
        &'a self,
        reader: &'a mut Reader<'_>,
        settings: &'a Self::Settings,
        load_context: &'a mut LoadContext<'_>
    ) -> BoxedFuture<'a, Result<Self::Asset, Self::Error>>;
    fn extensions(&self) -> &[&str];
}

Loads an Asset from a given byte Reader. This can accept AssetLoader::Settings, which configure how the Asset should be loaded.

Required Associated Types

type Asset: Asset

The top level Asset loaded by this AssetLoader.

type Settings: Settings + Default + Serialize + for<'a> Deserialize<'a>

The settings type used by this AssetLoader.

type Error: Into<Box<dyn Error + Send + Sync + 'static>>

The type of error which could be encountered by this loader.

Required Methods

fn load<'a>( &'a self, reader: &'a mut Reader<'_>, settings: &'a Self::Settings, load_context: &'a mut LoadContext<'_> ) -> BoxedFuture<'a, Result<Self::Asset, Self::Error>>

Asynchronously loads AssetLoader::Asset (and any other labeled assets) from the bytes provided by Reader.

fn extensions(&self) -> &[&str]

Returns a list of extensions supported by this asset loader, without the preceding dot.